Implement subscription on channel

+++ src/main/java/org/smirnowku/notifier/service/ChatService.java
@@ -6,5 +6,5 @@ public interface ChatService {
 
     Chat create(Chat chat);
 
-    Chat getByTelegramId(int telegramId);
+    Chat getByTelegramId(long telegramId);
 }
